// src/lib/api/types/addCategoryTranslationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface AddCategoryTranslationRequest {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** Translated description */
  description?: string;
  /**
   * Language code
   * @minLength 2
   * @maxLength 10
   */
  language: string;
  /**
   * Translated name
   * @minLength 1
   * @maxLength 200
   */
  name: string;
}