// src/lib/api/types/categoryDetailResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */
import type { CategoryTranslationItem } from './categoryTranslationItem';

export interface CategoryDetailResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** Icon name */
  icon?: string;
  /** Category ID */
  id: string;
  /** Active flag */
  isActive: boolean;
  /** Category name */
  name: string;
  /** Parent category ID */
  parentCategoryId?: string;
  /** Category slug */
  slug: string;
  /** Display order */
  sortOrder: number;
  /**
   * Category translations
   * @nullable
   */
  translations?: CategoryTranslationItem[] | null;
}