// src/lib/api/types/categorySummaryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface CategorySummaryResponse {
  /** Icon name */
  icon?: string;
  /** Category ID */
  id: string;
  /** Active flag */
  isActive: boolean;
  /** Category name */
  name: string;
  /** Category slug */
  slug: string;
  /** Display order */
  sortOrder: number;
}