// src/lib/api/types/createLibraryCategoryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface CreateLibraryCategoryRequest {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** Icon name */
  icon?: string;
  /**
   * Category name
   * @minLength 1
   * @maxLength 200
   */
  name: string;
  /** Parent category ID */
  parentCategoryId?: string;
  /**
   * URL-friendly identifier
   * @minLength 1
   * @maxLength 200
   */
  slug: string;
  /** Display order */
  sortOrder: number;
}