// src/lib/api/types/downloadLogListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */
import type { DownloadLogResponse } from './downloadLogResponse';

export interface DownloadLogListResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /**
   * Download logs
   * @nullable
   */
  data: DownloadLogResponse[] | null;
  /** Current page */
  page: number;
  /** Items per page */
  pageSize: number;
  /** Total count */
  total: number;
  /** Total pages */
  totalPages: number;
}