// src/lib/api/types/libraryGetDownloadLogsParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export type LibraryGetDownloadLogsParams = {
/**
 * Filter by template group ID
 */
groupId?: string;
/**
 * Page number
 */
page?: number;
/**
 * Items per page
 * @maximum 100
 */
pageSize?: number;
};